# 003 Code/Aiary-Backend-Public-main/swagger-api.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "OpenAPI definition",
    "version": "v0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Generated server url"
    }
  ],
...
  "paths": {
    "/diaries": {
      "get": {
        "tags": [
          "Diary"
        ],
        "summary": "일기 목록 조회",
        "description": "특정 스레드의 일기 목록을 페이지네이션으로 조회합니다",
        "operationId": "list",
        "parameters": [
          {
            "name": "threadId",
            "in": "query",
            "description": "스레드 ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "조회할 항목 수 (최대 100)",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 20
            },
            "example": 20
          },
          {
            "name": "cursor",
            "in": "query",
            "description": "페이지네이션 커서 (이전 응답의 nextCursor 값)",
            "required": false,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "일기 목록 조회 성공",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "object"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Diary"
        ],
        "summary": "일기 생성",
        "description": "스레드의 메시지를 기반으로 일기를 생성합니다",
        "operationId": "create",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateDiaryRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "404": {
            "description": "스레드를 찾을 수 없음",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "object"
                  }
                }
              }
            }
          },
          "400": {
            "description": "잘못된 요청",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "object"
                  }
                }
              }
            }
          },
          "200": {
            "description": "일기 생성 성공",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "object"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/diaries/{diaryId}": {
      "get": {
        "tags": [
          "Diary"
        ],
        "summary": "일기 조회",
        "description": "특정 일기의 상세 정보를 조회합니다",
        "operationId": "get",
        "parameters": [
          {
            "name": "diaryId",
            "in": "path",
            "description": "일기 ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "일기 조회 성공",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Diary"
                }
              }
            }
          },
          "404": {
            "description": "일기를 찾을 수 없음",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Diary"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "CreateDiaryRequest": {
        "type": "object",
        "properties": {
          "threadId": {
            "type": "string",
            "format": "uuid"
          },
          "title": {
            "type": "string"
          },
          "mood": {
            "type": "string"
          }
        }
      },
...
      "Diary": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "threadId": {
            "type": "string",
            "format": "uuid"
          },
          "title": {
            "type": "string"
          },
          "mood": {
            "type": "string"
          },
          "content": {
            "type": "string"
          },
          "summary": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
    }
  }
}